requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
anyio==4.6.2.post1
APScheduler==3.10.4
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
multidict==6.1.0
propcache==0.2.0
python-telegram-bot==21.6
pytz==2024.2
six==1.16.0
sniffio==1.3.1
tzdata==2024.2
tzlocal==5.2
yarl==1.16.0
